Move validation to core widgets
